-started setting up vmcb
-added  __sgdt intrin
-added some comments
-added (i/g)dtr struct

+++ svm_intercept_code.h
@@ -3,7 +3,7 @@
 
 //This is for SEV-ES
 //This should be all... but i used chatgippity sooo
-enum svm_intercept_code : uint64_t {
+enum svm_intercept_code : long long {
     // CR read/write operations
     VMEXIT_CR0_READ = 0x0, // Read of CR 0 through 15, respectively.
     VMEXIT_CR1_READ = 0x1,
